Add support for inverted text drawing
It was already possibly to draw text in white (kDrawModeFillWhite) by
setting `color` to white on the layer. However, this doesn't work well
for fonts which contain both black and white pixels.

This changes adds an optional `invertTextColor` parameter which can be
used to draw text inverted (kDrawModeInverted). Additionally, both
`invertTextColor` and `color` can be supplied for the layer, or as a
default for all layers on the Panel.