Support specifying custom transition duration and ease for each sequence
Automatic transitions previously used a default duration and ease. These
may now be overridden independently for each sequence in the comic with
the `transitionDuration` and `transitionEase` settings. If omitted, the
previous default behavior is retained.

NOTE: It might be interesting to support custom transitions for individual
panels, however that would entail a fair bit more complexity as one would
likley need to adjust behaviors for in vs. out transitions and look up the
settings for the next or previous panel before transitioning.