Avoid crash when cutscenes have different background colors
Sequence transitions cross fade when they have different background
colors. However, when using Panels for cutscenes this can cause a
crash because it attempts to set up a transition using an animator
that doesn't exist, since the sequence from the previous cutscene was
not active immediately before the new cutscene begins. To avoid this,
we just set the value which triggers the cross fade to `nil` when
a cutscene ends, as no transition is needed.